test(building): mirror source layout under tests/building/
tests/high_rise/ was the stale pre-rename name (the module became
cfdmod.building); its files tested cfdmod/building/* from a directory that
no longer matched any source package, while tests/building/ already held the
comfort/loadcases/peaks/modes tests. Consolidate:

- test_high_rise.py split into test_case.py (BuildingCase + case_data
  loading -> case.py) and test_pressure.py (Cp + per-floor Cf/Cm +
  per_floor_loads -> pressure.py)
- test_fanout.py, test_perf_memory.py moved verbatim
- tests/high_rise/ removed

Same 21 items collect (16 from the split + 4 fanout + 1 perf); path
resolution is unchanged (parents[2] is still the repo root at the same
directory depth).
